+++ index.vue
@@ -26,7 +26,7 @@
           <kmnotes  class="test" v-model="notes"></kmnotes>  
         </div>
         <div class="col-6">
-          <h3>Link</h3>{{ data }} 
+          <h3>Link</h3>
             <!-- {{ relevantDocuments }}  -->
             <!-- <table class="table table-striped" v-if="relevantDocuments?.length">
               <tr  v-for="(item, index) in relevantDocuments" :key="index">
@@ -34,8 +34,12 @@
                   <td style="width:10px;"><span class="btn" @click="remove(index)"><i class="fa fa-trash-o icon"></i></span></td>
               </tr>
           </table>    -->
-         
-          <kmlink  class="test" v-model="data"></kmlink>  
+          <p>relevant info is : </p>
+          <p>{{ relevantInfomation }}</p>
+          <p>relevant documents is : </p>
+          <p>{{ relevantDocuments }}</p>
+          <!-- <kmlink  class="test" v-model="data"></kmlink>   -->
+          <kmlink  class="test" v-model:relevantInfomation="relevantInfomation" v-model:relevantDocuments = "relevantDocuments"></kmlink>  
         </div>
       </div>
       <div class="row mt-5">
@@ -62,7 +66,7 @@
   import CopCountDown from './src/components/countdown/cop-count-down.vue';
   import dateSelector from  './src/components/inputs/dateSelector.vue'
   import kmnotes from  './src/components/inputs/notes.vue'
-  import kmlink from  './src/components/controls/link.vue'
+  import kmlink from  './src/components/controls/km-link.vue'
   import { ref ,provide} from 'vue';
 
   provide('auth', {
@@ -102,9 +106,9 @@
 
   const notes = ref("[\"[ Yuan Li | Feb 29, 2024 3:04:00 PM ] - this is first notes\",\"[ Yuan Li | Feb 29, 2024 5:26:42 PM ] - this is second notes\"]")
  
-  const relevantInfomation = {"en": "<div><!--block-->this is some info</div>"}
+  const relevantInfomation = ref({"en": "<div><!--block-->this is some info</div>"})
 
-  const relevantDocuments =[
+  const relevantDocuments =ref([
       {
         "url": "https://www.dsfgsdfgsdfg",
         "name": "dsgsdfgsdfgsdfg",
@@ -124,8 +128,8 @@
         "url": "/api/v2013/documents/DFF1283A-411A-75CF-B678-0A6EA5696070/attachments/614775/TEST.txt",
         "name": "TEST.txt",
         "language": "ar"
-      }]
-    const data = ref({relevantInfomation,relevantDocuments})
+      }])
+    //const data = ref({relevantInfomation,relevantDocuments})
    
 </script>
 
@@ -142,4 +146,4 @@
   filter: drop-shadow(0 0 2em #42b883aa);
 }
 </style>
-  
+  ./src/components/controls/km-link.vue
